+++ src/Ouzo/Inject/Injection/Factory.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Ouzo\Injection;
+
+interface Factory
+{
+    public function create();
+}