+++ src/Ouzo/Inject/Injection/InstanceRepository.php
@@ -3,6 +3,7 @@
  * Copyright (c) Ouzo contributors, http://ouzoframework.org
  * This file is made available under the MIT License (view the LICENSE file for more information).
  */
+
 namespace Ouzo\Injection;
 
 use BadMethodCallException;
@@ -12,11 +13,23 @@ class InstanceRepository
     /** @var object[] */
     private $instances = [];
 
+    /** @var Bindings */
+    private $bindings;
+
+    /**
+     * @param Bindings $bindings
+     */
+    public function __construct(Bindings $bindings)
+    {
+        $this->bindings = $bindings;
+    }
+
     /**
      * @param InstanceFactory $factory
      * @param Binder $binder
      * @return object
      * @throws BadMethodCallException
+     * @throws InjectorException
      */
     public function getInstance(InstanceFactory $factory, Binder $binder)
     {
@@ -24,6 +37,12 @@ class InstanceRepository
         if ($instance) {
             return $instance;
         }
+
+        $factoryClassName = $binder->getFactoryClassName();
+        if ($factoryClassName) {
+            return $this->createInstanceThroughFactory($factory, $factoryClassName);
+        }
+
         $className = $binder->getBoundClassName() ?: $binder->getClassName();
         $scope = $binder->getScope();
         if ($scope == Scope::SINGLETON) {
@@ -48,5 +67,21 @@ class InstanceRepository
         $instance = $factory->createInstance($this, $className);
         $this->instances[$className] = $instance;
         return $instance;
+    }
+
+    /**
+     * @param InstanceFactory $factory
+     * @param $factoryClassName
+     * @return mixed
+     * @throws InjectorException
+     */
+    private function createInstanceThroughFactory(InstanceFactory $factory, $factoryClassName)
+    {
+        if (!in_array(Factory::class, class_implements($factoryClassName))) {
+            throw new InjectorException("Factory class $factoryClassName does not implemented \Ouzo\Injection\Factory interface.");
+        }
+        $factoryBinder = $this->bindings->getBinder($factoryClassName);
+        $factoryObject = $this->getInstance($factory, $factoryBinder);
+        return $factoryObject->create();
     }
 }