+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Enhancements:
 * [ORM] Implemented upsert functionality (Model::createOrUpdate).
 * [DI] Added constructor injection for arguments with types defined (issue #265).
 * [DI] Implemented injection of private fields for parent class.
+* [DI] Implemented injection through factory class.
 * [Utilities] Added equalsIgnoreCase to Functions and FluentFunctions (issue #263).
 * [Utilities] Fixed Clock to support DST changes when adding hours, minutes or seconds.